moved firebase into config
