+++ pnaa/src/backend/auth/SignIn.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
 import { signInWithEmailAndPassword } from "firebase/auth";
-import { auth } from '../../firebase-config';
+import { auth } from '../../config/firebase-config';
 import { signInWithGoogle } from './AuthService';
 import { doc, getDoc, getFirestore } from "firebase/firestore";
